requirements.txt
Pillow
PyYAML
tk
